services/earthquake-data-aggregator-service/requirements.txt
certifi==2024.2.2
gevent==24.2.1
greenlet==3.0.3
websocket==0.2.1
zope.event==5.0
zope.interface==6.3
